More updates for using dagger to build the env.
This does not work yet.
Right now the socket does not appear to be forwarded into the container,
which is blocking forward progress.

This also changes the strategy from what run.sh does.
If all goes well, we will no longer need SSH or the agent to inject ssh
keys.
Instead we'll forward a socket fd into the VM and have dockerd use that.

Again... this is all incomplete and does not yet work.