Major refactor: Move code out of main and into lib
This splits up a lot of functionality.
The container entrypoint code is also moved to a separate binary.

The binaries that exist in the VM and the container are now built from
code that is embedded into the runner binary.
So the sources in cmd/init and cmd/entrypoint are embedded now and built
by dagger when creating the VM and container images respectively.

This change makes it possible to import this repo as a library and
`runner` is just an example of how to build something.
There is sitll some refactoring to do on `runner` to slim it down more,
but this is a good start.